fix(evalharness): survive scoring errors, drop dead score field, atomic chaos report; rename package
Address review feedback on the harness PR (#120):

- Rename the package devops_bench/harness -> devops_bench/evalharness (and
  tests/unit/harness -> tests/unit/evalharness) so it is not confused with the
  agent harness; update all imports, logger channels, and migration-doc paths.
  Public classes (Harness, DefaultHarness, ResultReporter, ScenarioManager)
  are unchanged.
- run() now wraps _score + the post-score rewrite in try/except: a judge or
  config failure (e.g. get_judge_model()) no longer escapes run() and discard
  an otherwise successful execution pass, whose raw results are already on disk.
- Drop the scalar `score` field from every record (and from _RECORD_KEYS): it
  was seeded to 0 and never written, so every record shipped score: 0. The
  per-metric `scores` map is the source of truth; tests updated accordingly.
- _drain_scenario deep-copies the chaos report on the join-timeout branch
  before stamping status, since the daemon thread may still be mid-write and a
  shallow dict() could capture a torn nested structure.
